Make VortexReadAt return static futures
