+++ vortex-datafusion/src/persistent/cache.rs
@@ -12,9 +12,8 @@ use object_store::{ObjectMeta, ObjectStore};
 use vortex::buffer::ByteBuffer;
 use vortex::dtype::DType;
 use vortex::error::{VortexError, VortexResult, vortex_err};
-use vortex::file::segments::SegmentCache;
 use vortex::file::{Footer, SegmentSpec, VortexFile, VortexOpenOptions};
-use vortex::layout::segments::SegmentId;
+use vortex::layout::segments::{SegmentCache, SegmentId};
 use vortex::session::VortexSession;
 use vortex::stats::{Precision, Stat};
 use vortex::utils::aliases::DefaultHashBuilder;
