+++ vortex-layout/src/segments/mod.rs
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
+mod cache;
 mod events;
+mod shared;
 mod sink;
 mod source;
 
@@ -11,7 +13,9 @@ mod test;
 use std::fmt::Display;
 use std::ops::Deref;
 
+pub use cache::*;
 pub use events::*;
+pub use shared::*;
 pub use sink::*;
 pub use source::*;
 #[cfg(any(test, feature = "test-harness"))]
