+++ vortex-layout/src/segments/shared.rs
@@ -1,0 +1,138 @@
+use std::sync::Arc;
+
+use futures::future::{BoxFuture, WeakShared};
+use futures::{FutureExt, TryFutureExt};
+use vortex_buffer::ByteBuffer;
+use vortex_error::{SharedVortexResult, VortexError, VortexExpect};
+use vortex_utils::aliases::dash_map::{DashMap, Entry};
+
+use crate::segments::{SegmentFuture, SegmentId, SegmentSource};
+
+/// A [`SegmentSource`] that allows multiple requesters to await the same underlying segment
+/// request.
+pub struct SharedSegmentSource<S> {
+    inner: S,
+    in_flight: DashMap<SegmentId, WeakShared<SharedSegmentFuture>>,
+}
+
+type SharedSegmentFuture = BoxFuture<'static, SharedVortexResult<ByteBuffer>>;
+
+impl<S: SegmentSource> SharedSegmentSource<S> {
+    /// Create a new `SharedSegmentSource` wrapping the provided inner source.
+    pub fn new(inner: S) -> Self {
+        Self {
+            inner,
+            in_flight: DashMap::default(),
+        }
+    }
+}
+
+impl<S: SegmentSource> SegmentSource for SharedSegmentSource<S> {
+    fn request(&self, id: SegmentId) -> SegmentFuture {
+        loop {
+            match self.in_flight.entry(id) {
+                Entry::Occupied(e) => {
+                    if let Some(shared_future) = e.get().upgrade() {
+                        return shared_future.map_err(VortexError::from).boxed();
+                    } else {
+                        // The future has been dropped, remove the entry and try again.
+                        e.remove();
+                    }
+                }
+                Entry::Vacant(e) => {
+                    let future = self.inner.request(id).map_err(Arc::new).boxed().shared();
+                    e.insert(
+                        future
+                            .downgrade()
+                            .vortex_expect("just created, cannot be polled to completion"),
+                    );
+                    return future.map_err(VortexError::from).boxed();
+                }
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::atomic::{AtomicUsize, Ordering};
+
+    use vortex_buffer::ByteBuffer;
+
+    use super::*;
+    use crate::segments::{SegmentSink, TestSegments};
+    use crate::sequence::SequenceId;
+
+    // Custom source that tracks how many times a segment is requested
+    #[derive(Default, Clone)]
+    struct CountingSegmentSource {
+        segments: TestSegments,
+        request_count: Arc<AtomicUsize>,
+    }
+
+    impl SegmentSource for CountingSegmentSource {
+        fn request(&self, id: SegmentId) -> SegmentFuture {
+            self.request_count.fetch_add(1, Ordering::SeqCst);
+            self.segments.request(id)
+        }
+    }
+
+    #[tokio::test]
+    async fn test_shared_source_deduplicates_concurrent_requests() {
+        let source = CountingSegmentSource::default();
+
+        // Add a segment to the test source
+        let data = ByteBuffer::from(vec![1, 2, 3, 4]);
+        let seq_id = SequenceId::root().downgrade();
+        source
+            .segments
+            .write(seq_id, vec![data.clone()])
+            .await
+            .unwrap();
+
+        let shared_source = SharedSegmentSource::new(source.clone());
+
+        // Request the same segment twice concurrently
+        let id = SegmentId::from(0);
+        let future1 = shared_source.request(id);
+        let future2 = shared_source.request(id);
+
+        // Both futures should resolve to the same data
+        let (result1, result2) = futures::join!(future1, future2);
+        assert_eq!(result1.unwrap(), data);
+        assert_eq!(result2.unwrap(), data);
+
+        // The inner source should have been called only once
+        assert_eq!(source.request_count.load(Ordering::Relaxed), 1);
+    }
+
+    #[tokio::test]
+    async fn test_shared_source_handles_dropped_futures() {
+        let source = CountingSegmentSource::default();
+
+        // Add a segment
+        let data = ByteBuffer::from(vec![5, 6, 7, 8]);
+        let seq_id = SequenceId::root().downgrade();
+        source
+            .segments
+            .write(seq_id, vec![data.clone()])
+            .await
+            .unwrap();
+
+        let shared_source = SharedSegmentSource::new(source.clone());
+        let id = SegmentId::from(0);
+
+        // Create and immediately drop a future
+        {
+            let _future = shared_source.request(id);
+            // Future is dropped here
+        }
+
+        // A new request should still work correctly
+        let result = shared_source.request(id).await;
+        assert_eq!(result.unwrap(), data);
+
+        // Should have made 2 requests since the first was dropped before completion
+        assert_eq!(source.request_count.load(Ordering::Relaxed), 2);
+    }
+}
